+++ RefRed/interfaces/mytablewidget.py
@@ -1,5 +1,5 @@
 # standard imports
-from enum import Enum
+from enum import IntEnum
 
 # third party imports
 from qtpy import QtWidgets
@@ -29,7 +29,7 @@ class MyTableWidget(QtWidgets.QTableWidget):
         self.parent = ui_parent
 
 
-class ReductionTableColumIndex(Enum):
+class ReductionTableColumnIndex(IntEnum):
     """
     Enumeration class associating a column index to a word easy to understand what the column index is for
     """
@@ -42,7 +42,8 @@ class ReductionTableColumIndex(Enum):
     LAMBDA_MAX = 5
     Q_MIN = 6
     Q_MAX = 7
-    COMMENTS = 8
+    CONST_Q_BINS = 8
+    COMMENTS = 9
 
     def __int__(self):
         return self.value