+++ test/unit/RefRed/reduction_table_handling/test_update_reduction_table.py
@@ -1,17 +1,11 @@
 from RefRed.main import MainGui
+from test.utilities import load_run_from_reduction_table
 
 # third party packages
 import pytest
 import unittest.mock as mock
 
 wait = 1000
-
-
-def load_run_from_reduction_table(window_main, row: int, col: int, run: str):
-    """Add run number in reduction table cell and press Enter to load a run"""
-    window_main.ui.reductionTable.setCurrentCell(row, col)
-    window_main.ui.reductionTable.currentItem().setText(run)
-    window_main.ui.table_reduction_cell_enter_pressed()
 
 
 @pytest.mark.parametrize("is_display_checked", [True, False])